Triagem/odq_recruta/requirements.txt
python-dotenv
msal
requests
pdfminer.six
python-docx
textract
rapidfuzz
scikit-learn
tenacity
tqdm
pytest
